% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_blood_cols.R
\name{add_blood_cols}
\alias{add_blood_cols}
\title{add_blood_cols}
\usage{
add_blood_cols(x, corlist)
}
\arguments{
\item{x}{data frame eqtl result}

\item{corlist}{output from correlate_blood_prostate}
}
\value{
x with two added columns - bloodPCC and bloodP
}
\description{
This function will add two columns to eqtl result
}
\examples{
load("data/res.Rdata")
load("data/blood_prostate_cor.Rdata")
res <- add_blood_cols(res, blood_prostate_cor)
}
\seealso{
correlate_blood_prostate
}
