% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gwas.R
\name{add_gwas}
\alias{add_gwas}
\title{add_gwas}
\usage{
add_gwas(x, file_gwas)
}
\arguments{
\item{x}{dataframe of eqtl result}

\item{file_gwas}{file to gwas information}
}
\value{
Will add two columns - gwasTrait and gwasP
}
\description{
This function will add GWAS info
}
\examples{
load("data/res.Rdata")
res <- add_siglabel_class(res)
}
