% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_siglabel_class.R
\name{add_siglabel_class}
\alias{add_siglabel_class}
\title{add_siglabel_class}
\usage{
add_siglabel_class(x, fdr = 0.1, p = 0.05)
}
\arguments{
\item{x}{dataframe of eqtl result}

\item{fdr}{fdr cutoff; default = 0.1}

\item{p}{p cutoff; default = 0.05}
}
\value{
Will add two columns - sigdelta and class
}
\description{
This function will label eqtl results for their significance or activator/insulator class
}
\examples{
load("data/res.Rdata")
res <- add_siglabel_class(res)
}
