% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snp_cpg_gene_pos.R
\name{add_snp_cpg_gene_pos}
\alias{add_snp_cpg_gene_pos}
\title{add_snp_cpg_gene_pos}
\usage{
add_snp_cpg_gene_pos(
  file_res,
  file_cpg_ctcf,
  file_snp_loci,
  file_atac_peak,
  file_tss_loci,
  name
)
}
\arguments{
\item{file_res}{path to me_eqtl_res.Rdata}

\item{file_cpg_ctcf}{file of significant correlated CpG and CTCF}

\item{file_snp_loci}{path to SNP bed file}

\item{file_atac_peak}{file of ATAC-seq peaks with SNPs}

\item{file_tss_loci}{path to TSS bed file}

\item{name}{prefix of output files}
}
\description{
This function will add POS information of SNP/ATCA, CpG/CTCF and genes/TSS
And split the res to two parts sign of cor(CpG, CTCF)
}
