% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{corr_dotmap}
\alias{corr_dotmap}
\title{corr_dotmap
Dotmap based on correlation coefficients and p values}
\usage{
corr_dotmap(cor_r, cor_p, name)
}
\arguments{
\item{cor_r}{A vector or matrix of correltaion coefficient}

\item{cor_p}{A vector or matirx of correlation test p values}

\item{r_cut, p_cut}{cutoffs for the significant correlations}
}
\description{
corr_dotmap
Dotmap based on correlation coefficients and p values
}
\examples{

}
