% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{corr_scatter}
\alias{corr_scatter}
\title{corr_scatter
scatterplot for x and y with correlation and lm information
recommend for light x and y}
\usage{
corr_scatter(x, y, cor_m, x_lab, y_lab, title, fig_type)
}
\arguments{
\item{x, y}{A numeric vector}

\item{cor_m}{mehtod for correlation: "pearson", "kendall", "spearman"}

\item{x_lab}{The x axis lable}

\item{y_lab}{The y axis lable}

\item{title}{Name of output figure}

\item{fig_type}{The format of output figure, either "PNG/png" or "PDF/pdf"}
}
\description{
corr_scatter
scatterplot for x and y with correlation and lm information
recommend for light x and y
}
\examples{
x <- rnorm(30)
y <- rnorm(30)
corr_scatter(x, y, "pearson", "var_1", "var_2", "test_corr_scatter_plot", "PDF")
}
\author{
yzeng
}
