% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{corr_scatter_ci}
\alias{corr_scatter_ci}
\title{corr_scatter_ci
scatterplot for x and y with correlation and lm (with CI and Rugs);
recommend for median numbers of  x and y}
\usage{
corr_scatter_ci(x, y, cor_m, x_lab, y_lab, file_name)
}
\arguments{
\item{x, y}{A numeric vector}

\item{cor_m}{mehtod for correlation: "pearson", "kendall", "spearman"}

\item{x_lab}{The x axis lable}

\item{y_lab}{The y axis lable}

\item{file_name}{Name of output figure with format suffix}
}
\description{
corr_scatter_ci
scatterplot for x and y with correlation and lm (with CI and Rugs);
recommend for median numbers of  x and y
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
corr_scatter_ci(x, y, "pearson", "var_1", "var_2", "test_corr_scatter_with_CI.pdf")
}
\author{
yzeng
}
