% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{corr_scatter_hex}
\alias{corr_scatter_hex}
\title{corr_scatter_hex
Hexagonalfor x and y with correlation and lm
recommend for massive and compact x and y}
\usage{
corr_scatter_hex(x, y, cor_m, x_lab, y_lab, bin_cnt, file_name)
}
\arguments{
\item{x, y}{A numeric vector}

\item{cor_m}{mehtod for correlation: "pearson", "kendall", "spearman"}

\item{x_lab}{The x axis lable}

\item{y_lab}{The y axis lable}

\item{bin_cnt}{bins splict for x and y axis}

\item{file_name}{Name of output figure with format suffix}
}
\description{
corr_scatter_hex
Hexagonalfor x and y with correlation and lm
recommend for massive and compact x and y
}
\examples{
x <- rnorm(10000)
y <- rnorm(10000)
corr_scatter_hex(x, y, "pearson", "var_1", "var_2", 50, "test_corr_scatter_hex.pdf")
}
\author{
yzeng
}
