% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{corr_volcano}
\alias{corr_volcano}
\title{corr_volcano
Volcano plot for correlation coefficients and p values}
\usage{
corr_volcano(cor_r, cor_p, name, r_c = 0.3, p_c = 0.05)
}
\arguments{
\item{cor_r}{A numeric vector of correlation coefficients}

\item{cor_p}{A numeric vector of correlation p-values}

\item{name}{Name of output PNG figure}

\item{r_c}{The absuloate cutoff of correlation coefficient, default = 0.3}

\item{p_c}{The cor.test p-value cutoff, default = 0.05}
}
\description{
corr_volcano
Volcano plot for correlation coefficients and p values
}
\examples{
cor_r <- runif(1000, -1, 1)
cor_p <- runif(1000, 0, 1)
corr_volcano(cor_r, cor_p, "test_corr_volcano_plot.png")
}
\author{
yzeng
}
