% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{corr_volcano_ylim}
\alias{corr_volcano_ylim}
\title{corr_volcano_ylim
Volcano plot for correlation coefficients and p values. To deal with outliers
are with extreme p values, therefore if(y > ylim){y = ylim}}
\usage{
corr_volcano_ylim(cor_r, cor_p, y_lim, name, r_c = 0.3, p_c = 0.05)
}
\arguments{
\item{cor_r}{A numeric vector of correlation coefficients}

\item{cor_p}{A numeric vector of correlation p-values}

\item{y_lim}{The up limitation of the y}

\item{name}{Name of output PNG figure}

\item{r_c}{The absuloate cutoff of correlation coefficient, default = 0.3}

\item{p_c}{The cor.test p-value cutoff, default = 0.05}
}
\description{
corr_volcano_ylim
Volcano plot for correlation coefficients and p values. To deal with outliers
are with extreme p values, therefore if(y > ylim){y = ylim}
}
\examples{
cor_r <- runif(1000, -1, 1)
cor_p <- runif(1000, 0, 1)
corr_volcano_ylim(cor_r, cor_p, 2, "test_corr_volcano_ylim_plot.png")
}
\author{
yzeng
}
