% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_blood_prostate.R
\name{correlate_blood_prostate}
\alias{correlate_blood_prostate}
\title{correlate_blood_prostate}
\usage{
correlate_blood_prostate(blood_catalog, prostate_catalog, mdata)
}
\arguments{
\item{blood_catalog}{"blood/catalog.N.sorted.merged.bed"}

\item{prostate_catalog}{"encode_correlation/catalog.N.sorted.merged.all_samples.bed"}

\item{mdata}{dataframe of mdata}
}
\value{
a list of three elements - prostate mdata, blood mdata and correlation summary
}
\description{
This function will calculate the correlation between blood and prostate methylation
}
\examples{
file_blood_catalog <- "/Users/musaahmed/Google\ Drive/ctcf/blood/catalog.N.sorted.merged.bed"
file_prostate_catalog <- "/Users/musaahmed/Google\ Drive/ctcf/encode_correlation/catalog.N.sorted.merged.all_samples.bed"
file_mdata <- "/Users/musaahmed/Google\ Drive/ctcf/CPGEA_mdata.normal.txt"
mdata <- read.table(file_mdata, header=T, row.names = 1, sep="\t")
blood_prostate_cor <- correlate_blood_prostate(file_blood_catalog, file_prostate_catalog, mdata)
}
