% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_me_eQTLs.R
\name{.plot_cmb}
\alias{.plot_cmb}
\title{.plot_cmb}
\usage{
.plot_cmb(snp_geno, cpg_meth, gene_expr, effs, pvals, name)
}
\arguments{
\item{snp_geno}{vector of snp genotype}

\item{cpg_meth}{vector of cpg methylation}

\item{gene_expr}{vector of gene expression}

\item{effs}{effect sizes of me_eQTLs, eQTL and meQTL}

\item{name}{output plot file name prefix}
}
\description{
This function will draw combined plots inclduing me_eQTLs, eQTL and meQTL for each snp_cpg_gene combinations
}
