% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ctcf_in_prostate.R
\name{find_ctcf_in_prostate}
\alias{find_ctcf_in_prostate}
\title{find_ctcf_in_prostate}
\usage{
find_ctcf_in_prostate(x, ctcf_dir, ctcf_metadata)
}
\arguments{
\item{x}{dataframe of eqtl result}

\item{ctcf_dir}{string path to ctcf files}

\item{ctcf_metadata}{string path to ctcf metadata file}
}
\value{
will output a list of data.frames
}
\description{
This function will split the res data frame into a list of data frames.
The list will contain main result, cpg unique, snp unique, gene unique
}
