% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_reverse_eqtl.R
\name{find_reverse_eqtl}
\alias{find_reverse_eqtl}
\title{find_reverse_eqtl}
\usage{
find_reverse_eqtl(
  x,
  file_edata,
  file_gdata,
  file_mdata,
  probeid,
  snpid,
  methylation_status,
  cutAtMedian = T,
  writetable = F
)
}
\arguments{
\item{x}{eqtl dataframe, e.g., res$gene}

\item{file_edata}{path to edata file}

\item{file_gdata}{path to gdata file}

\item{file_mdata}{path to mdata file}

\item{probeid}{string probeid}

\item{snpid}{string snpid}

\item{methylation_status}{to calculate eQTLs in either High or Low methylation level of the probeid}

\item{cutAtMedian}{boolean to indicaite if methylation should be stratified at median, defailt T}

\item{writetable}{boolean to indicate if track files be saved or not}
}
\description{
This will find eQTLs for a subset with a particular methylation probe
}
