% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_histone_signals.R
\name{get_histone_signals}
\alias{get_histone_signals}
\alias{plot_histone_density}
\title{Extract histone ChIP-seq signals for probe regions}
\usage{
get_histone_signals(x)

plot_histone_density(x, histone_list, histone, cell, ...)
}
\arguments{
\item{x}{dataframe of eqtl result}

\item{histone_list}{output from get_histone_signals}

\item{histone}{which histone}

\item{cell}{which cell}
}
\value{
will output a list of data.frames
}
\description{
get_histone_signals
}
\details{
This function will extract histone levels in probe
}
