% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jci_links.R
\name{get_jci_links}
\alias{get_jci_links}
\title{get_jci_links}
\usage{
get_jci_links(x, excelfile)
}
\arguments{
\item{x}{dataframe of eqtl result}

\item{excelfile}{path to excel file}
}
\value{
will generate plot
}
\description{
This function will extract pol2 chia-pet from JCI paper
}
\examples{
load("data/res.Rdata")
res <- add_siglabel_class(res)
}
