% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ucsc_tracks.R
\name{get_ucsc_tracks}
\alias{get_ucsc_tracks}
\title{get_ucsc_tracks}
\usage{
get_ucsc_tracks(x)
}
\arguments{
\item{x}{dataframe of eqtl result}
}
\value{
will generate plot
}
\description{
This function will output tracks to load in UCSC browser
}
