% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_cpg_meth.R
\name{init_cpg_meth}
\alias{init_cpg_meth}
\title{This will filtering CpG methylation matrix and pull out corresponding CpG and CTCF bed file}
\usage{
init_cpg_meth(
  file_sample_id,
  file_cpg_meth,
  file_cpg_ctcf,
  meth_l = 0.25,
  meth_h = 0.75,
  name
)
}
\arguments{
\item{file_sample_id}{file of sample IDs}

\item{file_cpg_meth}{file of merged CpG sits methylation}

\item{file_cpg_ctcf}{file of significant correlated CpG and CTCF}

\item{meth_l}{cut-off of low methylation level,  default 0.25}

\item{meth_h}{cut-off of high methylation level,  default 0.75}

\item{name}{prefix of output files}
}
\description{
This will filtering CpG methylation matrix and pull out corresponding CpG and CTCF bed file
}
