% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_gene_expr.R
\name{init_gene_expr}
\alias{init_gene_expr}
\title{This will filtering expression matrix and pull out corresponding gene TSS bed file}
\usage{
init_gene_expr(
  file_sample_id,
  file_gene_expr,
  file_gene_gtf,
  expr_med = 1,
  name
)
}
\arguments{
\item{file_sample_id}{file of sample IDs}

\item{file_gene_expr}{file of gene expression matrix}

\item{file_gene_gtf}{file of gene annotation}

\item{expr_med}{cut-off of median gene expression,  default 1}

\item{name}{prefix of output files}
}
\description{
This will filtering expression matrix and pull out corresponding gene TSS bed file
}
