% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_snp_geno.R
\name{init_snp_geno}
\alias{init_snp_geno}
\title{This will filtering SNP genotype matrix and pull out corresponding SNP bed file}
\usage{
init_snp_geno(
  file_sample_id,
  file_snp_gt,
  file_snp_in_atac,
  maf_cut = 0.05,
  name
)
}
\arguments{
\item{file_sample_id}{file of sample IDs}

\item{file_snp_gt}{file of SNP genotype matrix}

\item{file_snp_in_atac}{bed file of SNPs located in ATAC distal peak region}

\item{maf_cut}{cut-off of MAF,  default 0.05}

\item{name}{prefix of output files}
}
\description{
This will filtering SNP genotype matrix and pull out corresponding SNP bed file
}
