% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_eqtl_res.R
\name{load_eqtl_res}
\alias{load_eqtl_res}
\title{load_eqtl_res}
\usage{
load_eqtl_res(x, cordat, atac)
}
\arguments{
\item{x}{eqtl result file}

\item{cordat}{"encode_correlation/methylation_correlated_with_ctcf.Padj_filtered.bed"}

\item{atac}{"CPGEA_gdata.ATAC.txt"}
}
\description{
This function will load eqtl result and retail one case per atac-ctcf-gene combination
}
\examples{
res<-load_eqtl_res("/Users/musaahmed/Google\ Drive/ctcf/out_all.noNA.normalized.txt", "/Users/musaahmed/Google\ Drive/ctcf/encode_correlation/methylation_correlated_with_ctcf.Padj_filtered.bed","/Users/musaahmed/Google\ Drive/ctcf/CPGEA_gdata.ATAC.txt")
}
\keyword{TCGA,metadata,boxplot}
