% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tracks.R
\name{make_tracks}
\alias{make_tracks}
\title{make_tracks}
\usage{
make_tracks(x)
}
\arguments{
\item{x}{dataframe of eqtl result}
}
\value{
will generate plot
}
\description{
This function will output tracks to load in UCSC browser
}
