% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dens_cdf.R
\name{plot_dens_cdf}
\alias{plot_dens_cdf}
\title{plot_dens_cdf}
\usage{
plot_dens_cdf(
  x,
  class_col = "class",
  y_col,
  method,
  nlog2 = F,
  bw = 10000,
  ...
)
}
\arguments{
\item{x}{dataframe of me_eQTL result}

\item{class_col}{string column name to used as separtors}

\item{y_col}{string column name (must be numeric column)}

\item{method}{what plot to draw, e.g.,  density, ecdf}

\item{nlog2}{boolean if y_col should be -log2 transformed. Default F}

\item{bw}{numeric density bw}
}
\value{
will generate plot
}
\description{
This function will generate densiy ot CDF plot from the eqtl result dataframe
}
