% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eqtl.R
\name{plot_eqtl}
\alias{plot_eqtl}
\title{plot_eqtl}
\usage{
plot_eqtl(x, snpid, probeid, geneid, plot_what = "all")
}
\arguments{
\item{x}{dataframe of eqtl result}

\item{snpid}{string snp}

\item{probeid}{string cpg id}

\item{geneid}{string gene}

\item{plot_what}{string what to plot, one of encode_correlation, all}
}
\value{
will generate plot
}
\description{
This function will draw plots for individual cases
}
\examples{
load("data/res.Rdata")
res <- add_siglabel_class(res)
}
