% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gwas.R
\name{plot_gwas}
\alias{plot_gwas}
\title{plot_gwas}
\usage{
plot_gwas(x, pdfname)
}
\arguments{
\item{x}{dataframe of eqtl result}

\item{pdfname}{path to pdf file to save}
}
\value{
will generate plot
}
\description{
This function will plot gwas data
}
