% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_me_eQTLs.R
\name{plot_me_eQTLs}
\alias{plot_me_eQTLs}
\title{plot_me_eQTLs}
\usage{
plot_me_eQTLs(
  snp_cpg_gene,
  file_edata,
  file_gdata,
  file_mdata,
  file_tss_loci,
  file_snp_loci,
  file_res,
  name
)
}
\arguments{
\item{snp_cpg_gene}{vector of snp_cpg_gene comination}

\item{file_edata}{path to edata file}

\item{file_gdata}{path to gdata file}

\item{file_mdata}{path to mdata file}

\item{file_tss_loci}{path to TSS bed file}

\item{file_snp_loci}{path to SNP bed file}

\item{file_res}{path to me_eqtl_res.Rdata}

\item{name}{output plot file name prefix}
}
\description{
This function will draw combined plots inclduing me_eQTLs, eQTL and meQTL for snp_cpg_gene combinations list
}
\examples{
snp_cpg_gene <- c("rs28520918_chr6:31402506_POU5F1", "rs1039750_chr11:94603277_SESN3")
plot_me_eQTLs(snp_cpg_gene, file_edata, file_gdata, file_mdata, file_snp_loci, file_res, "me_eQTLs_test")
}
