% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plot.R
\name{plot_plot}
\alias{plot_plot}
\alias{plot_blood_cor}
\title{plot_plot}
\usage{
plot_plot(x, class_col = "class", y_col, method, nlog2 = F, bw = 10000, ...)

plot_blood_cor(probeid, outfile)
}
\arguments{
\item{x}{dataframe of eqtl result}

\item{class_col}{string column name to used as separtors}

\item{y_col}{string column name (must be numeric column)}

\item{method}{what plot to draw, e.g., boxplot, density, ecdf}

\item{nlog2}{boolean if y_col should be -log2 transformed. Default F}

\item{bw}{numeric density bw}

\item{probeid}{string probeid}
}
\value{
will generate plot
}
\description{
This function will plot from the eqtl result dataframe
}
\examples{
load("data/res.Rdata")
res <- add_siglabel_class(res)
}
