% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{run_gsea}
\alias{run_gsea}
\alias{check_all_gsea}
\title{run_gsea}
\usage{
run_gsea(x, gmt, plot = F, siggenes, bggenes, pdffile, ...)

check_all_gsea(x)
}
\arguments{
\item{x}{eqtl dataframe, e.g., res$gene}

\item{gmt}{path to gmt file}

\item{plot}{boolean to plot barplot, default F}

\item{siggenes}{optional character vector of genes to test}

\item{bggenes}{optional character vector of genes to use as background}

\item{pdffile}{path to output pdf file}
}
\description{
This will estimate gsea
}
