% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_liftOver_wrapper.R
\name{run_liftOver_wrapper}
\alias{run_liftOver_wrapper}
\title{run_liftOver_wrapper}
\usage{
run_liftOver_wrapper(x, convertFrom = "hg19", chain, changeName = T)
}
\arguments{
\item{x}{GRanges object}

\item{convertFrom}{string to identify assembly version converting from}

\item{chain}{string link to chain file}

\item{changeName}{boolean, if T (default), the lifted over ranges will have chr:pos naming; if F, will retain original names as in x}
}
\description{
This is a wrapper for liftover function
}
