% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_matrixeqtl.R
\name{run_matrixeqtl}
\alias{run_matrixeqtl}
\title{find_eqtl_tests}
\usage{
run_matrixeqtl(
  file_edata,
  file_gdata,
  file_mdata,
  file_snp_loci,
  file_tss,
  file_atac,
  file_ctcf,
  geneSnpMaxDistance = 5e+05,
  cpgSnpMaxDistance = 5e+05
)
}
\arguments{
\item{file_edata}{path to edata file}

\item{file_gdata}{path to gdata file}

\item{file_mdata}{path to mdata file}

\item{file_snp_loci}{path to snp bed file}

\item{file_tss}{path to TSS bed file}

\item{file_atac}{path to file overlappign snp and ATAC}

\item{file_ctcf}{path to file overlapping cpgs and ctcf peaks in ENCODE}

\item{geneSnpMaxDistance}{max distance between snp and gene, default 500000}

\item{cpgSnpMaxDistance}{max distance between snp and probe, default 500000}

\item{x}{eqtl dataframe, e.g., res$gene}
}
\description{
This will run matrixeQTL
}
