% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_me_eQTL.R
\name{run_me_eQTL}
\alias{run_me_eQTL}
\title{This will run matrixeQTL for eQTL, meQTL and methylation-dependent eQTL}
\usage{
run_me_eQTL(
  file_edata,
  file_gdata,
  file_mdata,
  file_tss_loci,
  file_snp_loci,
  file_cpg_loci,
  geneSnpMaxDistance = 5e+05,
  cpgSnpMaxDistance = 5e+05,
  cluster_core = 2
)
}
\arguments{
\item{file_edata}{path to edata file}

\item{file_gdata}{path to gdata file}

\item{file_mdata}{path to mdata file}

\item{file_tss_loci}{path to TSS bed file}

\item{file_snp_loci}{path to SNP bed file}

\item{file_cpg_loci}{path to CpG bed file}

\item{geneSnpMaxDistance}{max distance between snp and gene, default 500000}

\item{cpgSnpMaxDistance}{max distance between snp and probe, default 500000}

\item{cluster_core}{doParallel core, default 2 for local}

\item{x}{eqtl dataframe, e.g., res$gene}
}
\description{
This will run matrixeQTL for eQTL, meQTL and methylation-dependent eQTL
}
