% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_eqtl_results.R
\name{split_eqtl_results}
\alias{split_eqtl_results}
\title{split_eqtl_results}
\usage{
split_eqtl_results(x)
}
\arguments{
\item{x}{dataframe of eqtl result}
}
\value{
will output a list of data.frames
}
\description{
This function will split the res data frame into a list of data frames.
The list will contain main result, cpg unique, snp unique, gene unique
}
